Change width of the img & txt pairs
Giving same width to both the image and the text part made them
same, giving .pair a max-width made the text stay in the same
confines as the images as well.

Giving each .pair margin of 16px also made them not spill out of
parents. Reduced gap on .info-containers (each flex item that
contains both the img and text).